again fixed auth system, there was really big problem with singletron. and wrote friends system, now you can add, remove and get list of friends

+++ ChatterServer/class/settings.cpp
@@ -2,7 +2,7 @@
 
 Settings::Settings() {}
 
-int Settings::getPort() const { return port; }
+quint16 Settings::getPort() const { return port; }
 
 void Settings::setPort(int value) { port = value; }
